Renamed folders and files to have a consistent naming convention
