+++ src/components/app/app.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import { HomeRoute } from '../../routes/home.route';
 import { DashboardRoute } from '../../routes/dashboard.route';
-import './App.css';
+import './app.css';
 import routes from '../../routes/route-constants';
 import PrivateRoute from '../../helpers/private.route';
