+++ src/routes/home.route.tsx
@@ -1,4 +1,4 @@
-import HomeContainer from '../components/Home/home.container';
+import HomeContainer from '../components/home/home.container';
 
 export const HomeRoute = () => {
     return <HomeContainer />;
